feat: Fase 0 bloque 2 — UI auth/onboarding + shell dashboard (shadcn)
- Tailwind + shadcn/ui (Button/Input/Label/Card) con tema violeta
- registro (argon2) + login (credenciales/Google) + onboarding (onboard_agency
  vía withUser/RLS) + guardas de acceso
- shell dashboard: sidebar, topbar, selector de agencia multi-tenant, signout
- seed: pnpm db:make-superadmin <email>
- fix: accounts.expires_at integer (esperado por adaptador Auth.js)
- next.config: outputFileTracingRoot al root del monorepo
- validado e2e en Postgres real + navegador: registro→onboarding→dashboard

+++ packages/db/migrations/0001_foundation.sql
@@ -59,7 +59,7 @@ create table if not exists accounts (
   provider_account_id text not null,
   refresh_token       text,
   access_token        text,
-  expires_at          bigint,
+  expires_at          integer,
   token_type          text,
   scope               text,
   id_token            text,
